some

+++ NEEngine/inc/FPSCounter.h
@@ -12,6 +12,6 @@ private:
 	DWORD m_startTime;
 
 	size_t m_totalFrameCount = 0;
-	double m_totalTime = 0.;
+	size_t m_totalTime = 0.;
 };
